land: power: Account for activity launch end hint
Currently we unconditionally boost as long as POWER_HINT_LAUNCH is
received. However, the framework sends the powerhint for when launch
starts and when launch finishes [1], and thereby causing the boosting
to last twice as long as we wanted.

1) Check whether the hint is a start or finish hint
2) Make use of launch finish hint to signal when to release the perflock
and don't rely on fixed duration boosting which has two shortcomings:
 * perflock can be held for too long, causing the CPU freq and other resources
   to be boosted for too long even if the application has finished launching,
   resulting in excessive battery drainage
 * perflock can be held for not long enough if it takes more than the timeout
   time to finish launch

[1]: https://github.com/LineageOS/android_frameworks_base/commit/65c7a15833b61bf2a3d193992326a68716f0aca4

Reference: Wahoo power-libperfmgr

Change-Id: I405cc453c5f58d9fb2583b9c6017f3964a0ce024